+++ changelog.md
@@ -3,3 +3,9 @@
 
 ## Changes
 - Update Arch Loom to 1.17
+
+## Fixes
+- Fix crash during fluid tank fill/drain
+- Fix item holder on neoforge doesn't register
+- Fix fluid entity holder identifier
+- Fix crash with energy on neoforge